% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccuracyStdErrorEstimation.R
\name{AccuracyStdErrorEstimation}
\alias{AccuracyStdErrorEstimation}
\title{Component-wise minimum variance semi-supervised regression.}
\usage{
AccuracyStdErrorEstimation(
  basis_labeled,
  basis_unlabeled,
  X_labeled,
  X_unlabeled,
  y,
  samp_prob,
  min_var_weight,
  beta_sl,
  beta_ssl,
  resids_beta_sl,
  resids_beta_imp,
  inverse_information,
  num_resamples = 500,
  threshold = 0.5
)
}
\arguments{
\item{X_labeled}{Covariate matrix for labeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}

\item{beta_SSL}{Semi-supervised regression coefficient vector.}

\item{beta_SL}{Supervised regression coefficient vector.}

\item{resids_imp}{Residuals from the imputation model.}

\item{resids_beta_SL}{Residuals from the supervised regression model.}

\item{epsilon}{Small offset to help with numerical stability.}
}
\value{
List with parameter and SE est as well as estimated min_var_weights.
}
\description{
Component-wise minimum variance semi-supervised regression.
}
