% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{CovariateGen}
\alias{CovariateGen}
\title{Generate data from a multivariate normal}
\usage{
CovariateGen(n, mu, sigma)
}
\arguments{
\item{n}{Size of data set.}

\item{mu}{Mean vector.}

\item{Sigma}{Covariance matrix.}
}
\value{
A matrix of multivariate normal data.
}
\description{
Generate data from a multivariate normal
}
