% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{InteractionBasis}
\alias{InteractionBasis}
\title{Basis of Interaction Terms}
\usage{
InteractionBasis(X)
}
\arguments{
\item{X}{Numeric matrix.}
}
\value{
Numeric matrix with basis containing interaction terms.
}
\description{
Basis of Interaction Terms
}
