% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervisedAccuracy.R
\name{SemiSupervisedAccuracy}
\alias{SemiSupervisedAccuracy}
\title{Semi-supervised error from a regression model.}
\usage{
SemiSupervisedAccuracy(
  basis_labeled,
  basis_unlabeled,
  X_labeled,
  X_unlabaled,
  y,
  beta_SSL,
  beta_imp,
  weight = NULL,
  threshold = NULL
)
}
\arguments{
\item{basis_labeled}{Basis matrix for labeled data set.}

\item{X_labeled}{Covariate matrix for labeled data set.}

\item{y}{Numeric outcome vector.}

\item{weight}{Numeric vector of resampling weights.}

\item{threshold}{Threshold for overall misclassification rate.}

\item{bais_unlabeled}{Basis matrix for unlabeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}
}
\value{
resamp_weightector containing regression coefficients.
}
\description{
Semi-supervised error from a regression model.
}
