% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervisedApparentAccuracy.R
\name{SemiSupervisedApparentAccuracy}
\alias{SemiSupervisedApparentAccuracy}
\title{Apparent estimates for brier score and overall misclassification rate.}
\usage{
SemiSupervisedApparentAccuracy(
  basis_labeled,
  basis_unlabeled,
  X_labeled,
  X_unlabaled,
  y,
  beta_SSL,
  beta_imp,
  samp_prob,
  resamp_weight = NULL,
  threshold = 0.5
)
}
\arguments{
\item{basis_labeled}{Basis matrix for labeled data set.}

\item{X_labeled}{Covariate matrix for labeled data set.}

\item{y}{Numeric outcome vector.}

\item{samp_prob}{Numeric vector of weights.}

\item{resamp_weight}{Numeric vector of resampling weights.}

\item{threshold}{Threshold for overall misclassification rate.}

\item{bais_unlabeled}{Basis matrix for unlabeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}
}
\value{
resamp_weightector containing regression coefficients.
}
\description{
Apparent estimates for brier score and overall misclassification rate.
}
