% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityRatioStdErrorEstimation.R,
%   R/StdErrorEstimation.R
\name{StdErrorEstimation}
\alias{StdErrorEstimation}
\title{Standard error estimation based on IF expansion for density ratio method.}
\usage{
StdErrorEstimation(X_labeled, X_unlabeled, y, beta, residual)

StdErrorEstimation(X_labeled, X_unlabeled, y, beta, residual)
}
\arguments{
\item{X_labeled}{Covariate matrix for labeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}

\item{y}{Numeric outcome vector.}

\item{beta}{Regression parameter estimate to compute the score function.}

\item{residual}{Estimated residual.}

\item{dr_projection}{Projection from density ratio (DR) method.}
}
\value{
Standard error estimate.

Standard error estimate.
}
\description{
Standard error estimation based on IF expansion for density ratio method.

Standard error estimation based on influence function expansion.
}
