% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{StratificationVar}
\alias{StratificationVar}
\title{Generate Stratification Variable}
\usage{
StratificationVar(covariates, num_strata = 2)
}
\arguments{
\item{covariates}{Covariate matrix.}

\item{num_strata}{Number of stratum.}
}
\value{
A stratification variable
}
\description{
Generate Stratification Variable
}
