% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{TwoWayInteractionBasis}
\alias{TwoWayInteractionBasis}
\title{Basis of Two-way Interaction Terms}
\usage{
TwoWayInteractionBasis(X, S)
}
\arguments{
\item{X}{Numeric matrix.}

\item{S}{Numeric matrix.}
}
\value{
Numeric matrix with basis containing interaction terms.
}
\description{
Basis of Two-way Interaction Terms
}
