% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccuracyStdErrorEstimation.R
\name{AccuracyStdErrorEstimation}
\alias{AccuracyStdErrorEstimation}
\title{Component-wise minimum variance semi-supervised regression.}
\usage{
AccuracyStdErrorEstimation(
  basis_labeled,
  basis_unlabeled,
  X_labeled,
  X_unlabeled,
  y,
  samp_prob,
  min_var_weight,
  beta_SL,
  beta_SSL,
  resids_beta_SL,
  resids_beta_imp,
  inverse_information,
  num_resamples = 500,
  threshold = 0.5
)
}
\arguments{
\item{basis_labeled}{Basis matrix for labeled data set.}

\item{X_labeled}{Covariate matrix for labeled data set.}

\item{X_unlabeled}{Covariate matrix for unlabeled data set.}

\item{y}{Numeric outcome vector.}

\item{samp_prob}{Numeric vector of weights.}

\item{min_var_weight}{Minimum variance weight for semi-supervised estimate.}

\item{beta_SL}{Supervised regression coefficient vector.}

\item{beta_SSL}{Semi-supervised regression coefficient vector.}

\item{resids_beta_SL}{Residuals from the supervised regression model.}

\item{resids_beta_imp}{Residuals from the imputation model.}

\item{inverse_information}{Inverse information  matric.}

\item{num_resamples}{Number of resamples.}

\item{threshold}{Threshold for over misclassification rate.}

\item{bais_unlabeled}{Basis matrix for unlabeled data set.}

\item{resids_beta_SSL}{Residuals from the semi-supervised regression model.}
}
\value{
Pertrubed estimates.
}
\description{
Component-wise minimum variance semi-supervised regression.
}
